requirements.txt
# Build tools
setuptools>=69.0.0
wheel>=0.42.0

# Core FastAPI and web server
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
python-multipart>=0.0.6,<0.1.0

# HTTP clients and web scraping
httpx>=0.25.0,<0.26.0
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
aiofiles>=23.2.0,<24.0.0
websockets>=12.0,<13.0

# OpenAI and AI
openai>=1.3.0,<2.0.0
tiktoken>=0.5.0,<1.0.0

# Database (PostgreSQL without asyncpg for now)
psycopg2-binary>=2.9.0,<3.0.0
sqlalchemy>=2.0.0,<2.1.0

# Data processing
numpy>=1.24.0,<2.0.0
pydantic>=2.5.0,<3.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
markdown>=3.5.0,<4.0.0

# Development dependencies
watchdog>=3.0.0,<4.0.0

# Optional: pgvector (install separately if needed)
# pgvector>=0.2.0,<0.3.0

# Optional: scikit-learn (install separately if needed)  
# scikit-learn>=1.3.0,<1.4.0

# Optional: MCP (install separately if needed)
# mcp>=1.0.0

# Optional: LangChain (install separately if needed)
# langchain>=0.1.0,<0.2.0
# langchain-openai>=0.1.0,<0.2.0
# langchain-community>=0.1.0,<0.2.0 